+++ src/lib.rs
@@ -29,6 +29,7 @@ mod driver;
 pub mod errors;
 mod gdal_major_object;
 mod metadata;
+pub mod programs;
 pub mod raster;
 pub mod spatial_ref;
 mod utils;
@@ -37,7 +38,8 @@ pub mod version;
 pub mod vsi;
 
 pub use dataset::{
-    Dataset, DatasetOptions, GdalOpenFlags, GeoTransform, LayerIterator, LayerOptions, Transaction,
+    Dataset, DatasetOptions, GdalOpenFlags, GeoTransform, GeoTransformEx, LayerIterator,
+    LayerOptions, Transaction,
 };
 pub use driver::Driver;
 pub use metadata::Metadata;